Hook up portfolio nav link click events and search input keyup handler
Hooked up portfolio nav links so that when one is clicked, it's highlighted and set to 'active' class, and all others are turned off. Portfolio cards are filtered based on the filter dataset of the clicked link. Portfolio search input is hooked up with a keyup listener that filters portfolio cards based on if the input text matches the card's filter dataset